+++ packages/cli/test/common/augmentation.d.ts
